+++ src/Base/Theme.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Termage\Base;
+
+abstract class Theme
+{
+
+}