+++ src/Components/Block.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Termage\Components;
+
+use Termage\Base\Element;
+
+final class Block extends Element
+{
+}