+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+<a name="0.2.0"></a>
+# [0.2.0](https://github.com/termage/termage) (2021-10-03)
+* Renderer engine imrovements.
+* New Emoji Emoji Component.
+* New Colors Class.
+
 <a name="0.1.0"></a>
-# [0.1.0](https://github.com/clirad/clirad) (2021-10-01)
+# [0.1.0](https://github.com/termage/termage) (2021-10-01)
 * Initial release